+++ lib/augusto/h_u4_completada/configuracion_cuenta/configuracion_cuenta_widget.dart
@@ -135,23 +135,15 @@ class _ConfiguracionCuentaWidgetState extends State<ConfiguracionCuentaWidget>
                     crossAxisAlignment: CrossAxisAlignment.center,
                     children: [
                       AuthUserStreamWidget(
-                        builder: (context) => InkWell(
-                          splashColor: Colors.transparent,
-                          focusColor: Colors.transparent,
-                          hoverColor: Colors.transparent,
-                          highlightColor: Colors.transparent,
-                          onTap: () async {},
-                          child: Text(
-                            currentUserDisplayName,
-                            style: FlutterFlowTheme.of(context)
-                                .headlineMedium
-                                .override(
-                                  fontFamily: 'Inter',
-                                  color:
-                                      FlutterFlowTheme.of(context).primaryText,
-                                  letterSpacing: 0.0,
-                                ),
-                          ),
+                        builder: (context) => Text(
+                          currentUserDisplayName,
+                          style: FlutterFlowTheme.of(context)
+                              .headlineMedium
+                              .override(
+                                fontFamily: 'Inter',
+                                color: FlutterFlowTheme.of(context).primaryText,
+                                letterSpacing: 0.0,
+                              ),
                         ),
                       ),
                     ].divide(const SizedBox(height: 8.0)),
