+++ lib/augusto/uh1completada/password_olvidado/password_olvidado_model.dart
@@ -1,8 +1,8 @@
 import '/flutter_flow/flutter_flow_util.dart';
-import 'forgot_widget.dart' show ForgotWidget;
+import 'password_olvidado_widget.dart' show PasswordOlvidadoWidget;
 import 'package:flutter/material.dart';
 
-class ForgotModel extends FlutterFlowModel<ForgotWidget> {
+class PasswordOlvidadoModel extends FlutterFlowModel<PasswordOlvidadoWidget> {
   ///  State fields for stateful widgets in this page.
 
   // State field(s) for emailAddress widget.
