--- lib/paginas_duplicadssonla_his4/menu/menu_model.dart
@@ -1,11 +1,0 @@
-import '/flutter_flow/flutter_flow_util.dart';
-import 'menu_widget.dart' show MenuWidget;
-import 'package:flutter/material.dart';
-
-class MenuModel extends FlutterFlowModel<MenuWidget> {
-  @override
-  void initState(BuildContext context) {}
-
-  @override
-  void dispose() {}
-}
