--- lib/paginas_duplicadssonla_his4/perfil_usuario/perfil_usuario_model.dart
@@ -1,11 +1,0 @@
-import '/flutter_flow/flutter_flow_util.dart';
-import 'perfil_usuario_widget.dart' show PerfilUsuarioWidget;
-import 'package:flutter/material.dart';
-
-class PerfilUsuarioModel extends FlutterFlowModel<PerfilUsuarioWidget> {
-  @override
-  void initState(BuildContext context) {}
-
-  @override
-  void dispose() {}
-}
